Convert ints to longs

+++ src/Cecs475.BoardGames.ComputerOpponent/MinimaxAi.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace Cecs475.BoardGames.ComputerOpponent {
 	internal struct MinimaxBestMove {
-		public int Weight { get; set; }
+		public long Weight { get; set; }
 		public IGameMove Move { get; set; }
 	}
 
@@ -19,12 +19,12 @@ namespace Cecs475.BoardGames.ComputerOpponent {
 
 		public IGameMove FindBestMove(IGameBoard b) {
 			return FindBestMove(b,
-				true ? int.MinValue : int.MaxValue,
-				true ? int.MaxValue : int.MinValue,
+				true ? long.MinValue : long.MaxValue,
+				true ? long.MaxValue : long.MinValue,
 				mMaxDepth).Move;
 		}
 
-		private static MinimaxBestMove FindBestMove(IGameBoard b, int alpha, int beta, int depthLeft) {
+		private static MinimaxBestMove FindBestMove(IGameBoard b, long alpha, long beta, int depthLeft) {
 			return new MinimaxBestMove() {
 				Move = null
 			};
